move droppedOptions to constants.js
